app/requirements.txt
streamlit
grpcio
grpcio-tools